pyproject.toml
[project]
name = "mlm-form"
version = "0.1.0"
description = "Runs a single page web app based on FastHTML to create Machine Learning Model metadata."
authors = [
    { name = "Ryan Avery", email = "ryanbarryavery@gmail.com" }
]
dependencies = [
    "python-fasthtml>=0.3.4",
    "stac-model>=0.2.0",
    "pydantic>=2.8.2",
    "pystac>=1.10.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"mlm-form" = "mlm_form:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mlm_form"]
